[#IOPID-1820] moved types and interfaces
