+++ apps/session-manager/src/types/fast-login.ts
@@ -1,3 +1,4 @@
+import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import { enumType } from "@pagopa/ts-commons/lib/types";
 import * as t from "io-ts";
 
@@ -10,3 +11,8 @@ export const LoginType = enumType<LoginTypeEnum>(LoginTypeEnum, "LoginType");
 
 export type AdditionalLoginPropsT = t.TypeOf<typeof AdditionalLoginProps>;
 export const AdditionalLoginProps = t.partial({ loginType: LoginType });
+
+export const FastLoginResponse = t.type({
+  token: NonEmptyString,
+});
+export type FastLoginResponse = t.TypeOf<typeof FastLoginResponse>;
